Finalize graph implementations
